fix: allow Gliederungadmins to edit photos
