+++ apps/api/src/services/person/person.router.ts
@@ -2,10 +2,9 @@
 import { mergeRouters } from '../../trpc.js'
 import { personAuthenticatedGetProcedure } from './personAuthenticatedGet.js'
 import { personGetProcedure } from './personGet.js'
-import { personGliederungPatchProcedure } from './personGliederungPatch.js'
 import { personCountProcedure, personListProcedure } from './personList.js'
 import { personVerwaltungCreateProcedure } from './personVerwaltungCreate.js'
-import { personVerwaltungPatchProcedure } from './personVerwaltungPatch.js'
+import { personVerwaltungPatchProcedure } from './personPatch.js'
 import { personVerwaltungRemoveProcedure } from './personVerwaltungRemove.js'
 // Import Routes here - do not delete this line
 
@@ -16,7 +15,6 @@ export const personRouter = mergeRouters(
   personCountProcedure.router,
   personGetProcedure.router,
   personVerwaltungPatchProcedure.router,
-  personVerwaltungRemoveProcedure.router,
-  personGliederungPatchProcedure.router
+  personVerwaltungRemoveProcedure.router
   // Add Routes here - do not delete this line
 )
