--- apps/api/src/services/person/personVerwaltungPatch.ts
@@ -1,33 +1,0 @@
-import { Role } from '@prisma/client'
-import z from 'zod'
-
-import prisma from '../../prisma.js'
-import { defineProtectedMutateProcedure } from '../../types/defineProcedure.js'
-
-import { getPersonCreateData, personSchemaOptional } from './schema/person.schema.js'
-
-export const personVerwaltungPatchProcedure = defineProtectedMutateProcedure({
-  key: 'verwaltungPatch',
-  roleIds: [Role.ADMIN],
-  inputSchema: z.strictObject({
-    id: z.number().int(),
-    data: personSchemaOptional,
-  }),
-  async handler({ input }) {
-    await prisma.notfallkontakt.deleteMany({
-      where: {
-        personId: input.id,
-      },
-    })
-
-    return prisma.person.update({
-      where: {
-        id: input.id,
-      },
-      data: await getPersonCreateData(input.data),
-      select: {
-        id: true,
-      },
-    })
-  },
-})
